feat: used vite in place of create-react-app
