+++ src/components/Post.js
@@ -1,0 +1,85 @@
+import React, { useEffect } from "react";
+import { Avatar } from "@mui/material";
+import { db } from "../lib/firebase";
+import firebase from "firebase/compat/app";
+
+function Post(prop) {
+  const { username, caption, imageUrl, avatar, postId, user } = prop;
+  const [comments, setComments] = React.useState([]);
+  const [comment, setComment] = React.useState("");
+
+  useEffect(() => {
+    let unsubscribe;
+    if (postId) {
+      unsubscribe = db
+        .collection("posts")
+        .doc(postId)
+        .collection("comments")
+        .orderBy("timestamp", "desc")
+        .onSnapshot((snapshot) => {
+          setComments(snapshot.docs.map((doc) => doc.data()));
+        });
+    }
+    return () => {
+      unsubscribe();
+    };
+  }, [postId]);
+
+  const postComment = (event) => {
+    event.preventDefault();
+    db.collection("posts").doc(postId).collection("comments").add({
+      text: comment,
+      username: user.displayName,
+      timestamp: firebase.firestore.FieldValue.serverTimestamp(),
+    });
+    setComment("");
+  };
+
+  return (
+    <div className="post">
+      <div className="post__header">
+        <Avatar className="post__avatar" alt={username} src={avatar} />
+        <h3 className="post__username">{username}</h3>
+      </div>
+
+      <div className="post__container">
+        <img className="post__img" src={imageUrl} alt="random sq" />
+
+        <h4 className="post__text">
+          <strong>{username} </strong>
+          {caption}
+        </h4>
+
+        <div className="post__comments">
+          {comments.map((userComment) => (
+            <p>
+              <strong>{userComment.username}</strong> {userComment.text}
+            </p>
+          ))}
+        </div>
+
+        {user && (
+          <form className="post__commentBox">
+            <input
+              className="post__input"
+              type="text"
+              placeholder="Add a comment..."
+              value={comment}
+              onChange={(e) => setComment(e.target.value)}
+            />
+            <button
+              className="post__button"
+              disabled={!comment}
+              type="submit"
+              onClick={postComment}
+            >
+              Post
+            </button>
+          </form>
+        )}
+      </div>
+    </div>
+  );
+}
+
+export default Post;
